server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors
PyJWT
Flask-TinyMCE
Flask-Mail
flask-socketio
